requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bleach==6.1.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==2.0.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.3.0
openai==1.7.2
packaging==23.2
proto-plus==1.23.0
protobuf==3.20.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
twilio==8.13.0
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
yarl==1.9.4
pymysql==1.1.1